backend/agents/parser.py
"""
Agent 1: URL Parser & Content Extractor

Responsibilities:
  1. Accept URL input from user
  2. Fetch webpage HTML
  3. Parse and extract: headline, body text, author, publish date, source domain
  4. Clean and structure content
  5. Output JSON with extracted data

Technologies: BeautifulSoup4, Newspaper3k, httpx
"""

import json
import logging
from datetime import datetime, timezone

import httpx
from newspaper import Article
from bs4 import BeautifulSoup

logger = logging.getLogger(__name__)


class ParserAgent:
    """Agent 1 — Extracts clean article content from a news URL."""

    AGENT_NAME = "parser"

    async def run(self, url: str) -> dict:
        """
        Fetch and parse the given news URL.

        Returns a dict matching the Agent 1 output schema:
        {
            "success": bool,
            "agent": "parser",
            "thought": str,
            "data": {
                "url": str,
                "headline": str,
                "body": str,
                "author": str,
                "publish_date": str,
                "domain": str,
                "word_count": int
            },
            "timestamp": str (ISO-8601)
        }
        """
        # TODO: Implement URL fetching, HTML parsing, and article extraction
        raise NotImplementedError("ParserAgent.run() is not yet implemented.")
