backend/requirements.txt
# ── Framework & Server ────────────────────────────────────
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# ── Async HTTP ────────────────────────────────────────────
httpx==0.26.0

# ── Web Scraping ──────────────────────────────────────────
newspaper3k==0.2.8
beautifulsoup4==4.12.3
lxml_html_clean==0.4.1

# ── LLM & AI ─────────────────────────────────────────────
groq==0.13.0

# ── Utilities ─────────────────────────────────────────────
python-dotenv==1.0.0
